requirements-pythonanywhere.txt
# Requirements for PythonAnywhere deployment
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Authlib==1.2.1
requests==2.31.0
python-dotenv==1.0.0
mutagen==1.47.0
Werkzeug==2.3.7
